SMART re-initialize in a smart way
Till now I created a basic structure for SMART project.
Now, it's time to re-organize, with new and real time datasets,
new and better API handling and Implement the project
in a smart way. We redefined the scope and requirements as well.

The project will now follow MLOps life cycle for better management
All the requirements and plan details are available
in README.md and SMART_phase_plan.md respectively.